TEEEM_DOCS/ARCHIVE/ml_service/requirements.txt
# Machine Learning Dependencies
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.3

# Database Connection
psycopg2-binary==2.9.9

# Task Queue & Scheduling
celery==5.3.6
redis==5.0.1

# Data Persistence
joblib==1.3.2

# Configuration & Environment
python-dotenv==1.0.0

# Logging & Monitoring
python-json-logger==2.0.7

# Optional: Time Series Analysis
statsmodels==0.14.1
